+++ lib/Output/OPDS_renderer.php
@@ -8,26 +8,23 @@
 
 namespace SebLucas\Cops\Output;
 
+use SebLucas\Cops\Input\Config;
+use SebLucas\Cops\Input\Request;
 use SebLucas\Cops\Model\EntryBook;
 use SebLucas\Cops\Model\Link;
 use SebLucas\Cops\Model\LinkFacet;
 use SebLucas\Cops\Model\LinkNavigation;
+use SebLucas\Cops\Output\Format;
 use SebLucas\Cops\Pages\Page;
 use XMLWriter;
 
-use function SebLucas\Cops\Language\localize;
-use function SebLucas\Cops\Request\addURLParameter;
-use function SebLucas\Cops\Request\getQueryString;
-use function SebLucas\Cops\Request\getURLParam;
-
-use const SebLucas\Cops\Config\COPS_DB_PARAM;
-use const SebLucas\Cops\Config\COPS_ENDPOINTS;
-
 class OPDSRenderer
 {
-    public static $endpoint = COPS_ENDPOINTS["feed"];
+    public static $endpoint = Config::ENDPOINT["feed"];
     private $xmlStream = null;
     private $updated = null;
+    /** @var Request */
+    protected $request;
 
     private function getUpdatedTime()
     {
@@ -47,9 +44,15 @@ class OPDSRenderer
         return $this->xmlStream;
     }
 
-    public function getOpenSearch()
+    /**
+     * Summary of getOpenSearch
+     * @param Request $request
+     * @return string
+     */
+    public function getOpenSearch($request)
     {
         global $config;
+        $database = $request->get('db');
         $xml = new XMLWriter();
         $xml->openMemory();
         $xml->setIndent(true);
@@ -77,9 +80,7 @@ class OPDSRenderer
         $xml->startElement("Url");
         $xml->writeAttribute("type", 'application/atom+xml');
         $urlparam = "?query={searchTerms}";
-        if (!is_null(getURLParam(COPS_DB_PARAM))) {
-            $urlparam = addURLParameter($urlparam, COPS_DB_PARAM, getURLParam(COPS_DB_PARAM));
-        }
+        $urlparam = Format::addDatabaseParam($urlparam, $database);
         $urlparam = str_replace("%7B", "{", $urlparam);
         $urlparam = str_replace("%7D", "}", $urlparam);
         $xml->writeAttribute("template", $config['cops_full_url'] . self::$endpoint . $urlparam);
@@ -93,9 +94,10 @@ class OPDSRenderer
         return $xml->outputMemory(true);
     }
 
-    private function startXmlDocument($page)
+    private function startXmlDocument($page, $request)
     {
         global $config;
+        $database = $request->get('db');
         self::getXmlStream()->startDocument('1.0', 'UTF-8');
         self::getXmlStream()->startElement("feed");
         self::getXmlStream()->writeAttribute("xmlns", "http://www.w3.org/2005/Atom");
@@ -114,8 +116,8 @@ class OPDSRenderer
         self::getXmlStream()->startElement("id");
         if ($page->idPage) {
             $idPage = $page->idPage;
-            if (!is_null(getURLParam(COPS_DB_PARAM))) {
-                $idPage = str_replace("cops:", "cops:" . getURLParam(COPS_DB_PARAM) . ":", $idPage);
+            if (!is_null($request->get('db'))) {
+                $idPage = str_replace("cops:", "cops:" . $request->get('db') . ":", $idPage);
             }
             self::getXmlStream()->text($idPage);
         } else {
@@ -141,28 +143,26 @@ class OPDSRenderer
         self::getXmlStream()->endElement();
         $link = new LinkNavigation("", "start", "Home");
         self::renderLink($link);
-        $link = new LinkNavigation("?" . getQueryString(), "self");
+        $link = new LinkNavigation("?" . $request->query(), "self");
         self::renderLink($link);
         $urlparam = "?";
-        if (!is_null(getURLParam(COPS_DB_PARAM))) {
-            $urlparam = addURLParameter($urlparam, COPS_DB_PARAM, getURLParam(COPS_DB_PARAM));
-        }
-        if ($config['cops_generate_invalid_opds_stream'] == 0 || preg_match("/(MantanoReader|FBReader)/", $_SERVER['HTTP_USER_AGENT'])) {
+        $urlparam = Format::addDatabaseParam($urlparam, $database);
+        if ($config['cops_generate_invalid_opds_stream'] == 0 || preg_match("/(MantanoReader|FBReader)/", $request->agent())) {
             // Good and compliant way of handling search
-            $urlparam = addURLParameter($urlparam, "page", Page::OPENSEARCH);
+            $urlparam = Format::addURLParam($urlparam, "page", Page::OPENSEARCH);
             $link = new Link(self::$endpoint . $urlparam, "application/opensearchdescription+xml", "search", "Search here");
         } else {
             // Bad way, will be removed when OPDS client are fixed
-            $urlparam = addURLParameter($urlparam, "query", "{searchTerms}");
+            $urlparam = Format::addURLParam($urlparam, "query", "{searchTerms}");
             $urlparam = str_replace("%7B", "{", $urlparam);
             $urlparam = str_replace("%7D", "}", $urlparam);
             $link = new Link($config['cops_full_url'] . self::$endpoint . $urlparam, "application/atom+xml", "search", "Search here");
         }
         self::renderLink($link);
         if ($page->containsBook() && !is_null($config['cops_books_filter']) && count($config['cops_books_filter']) > 0) {
-            $Urlfilter = getURLParam("tag", "");
+            $Urlfilter = $request->get("tag", "");
             foreach ($config['cops_books_filter'] as $lib => $filter) {
-                $link = new LinkFacet("?" . addURLParameter(getQueryString(), "tag", $filter), $lib, localize("tagword.title"), $filter == $Urlfilter);
+                $link = new LinkFacet("?" . Format::addURLParam($request->query(), "tag", $filter), $lib, localize("tagword.title"), $filter == $Urlfilter, $database);
                 self::renderLink($link);
             }
         }
@@ -267,10 +267,16 @@ class OPDSRenderer
         }
     }
 
-    public function render($page)
+    /**
+     * Summary of render
+     * @param mixed $page
+     * @param Request $request
+     * @return string
+     */
+    public function render($page, $request)
     {
         global $config;
-        self::startXmlDocument($page);
+        self::startXmlDocument($page, $request);
         if ($page->isPaginated()) {
             self::getXmlStream()->startElement("opensearch:totalResults");
             self::getXmlStream()->text($page->totalNumber);