+++ lib/Pages/PageAllAuthorsLetter.php
@@ -10,15 +10,12 @@ namespace SebLucas\Cops\Pages;
 
 use SebLucas\Cops\Calibre\Author;
 
-use function SebLucas\Cops\Language\localize;
-use function SebLucas\Cops\Language\str_format;
-
 class PageAllAuthorsLetter extends Page
 {
     public function InitializeContent()
     {
         $this->idPage = Author::getEntryIdByLetter($this->idGet);
-        $this->entryArray = Author::getAuthorsByStartingLetter($this->idGet);
+        $this->entryArray = Author::getAuthorsByStartingLetter($this->idGet, $this->getDatabaseId());
         $this->title = str_format(localize("splitByLetter.letter"), str_format(localize("authorword", count($this->entryArray)), count($this->entryArray)), $this->idGet);
     }
 }