+++ lib/Pages/PageAllPublishers.php
@@ -10,14 +10,12 @@ namespace SebLucas\Cops\Pages;
 
 use SebLucas\Cops\Calibre\Publisher;
 
-use function SebLucas\Cops\Language\localize;
-
 class PageAllPublishers extends Page
 {
     public function InitializeContent()
     {
         $this->title = localize("publishers.title");
-        $this->entryArray = Publisher::getAllPublishers();
+        $this->entryArray = Publisher::getAllPublishers($this->getDatabaseId());
         $this->idPage = Publisher::PAGE_ID;
     }
 }