+++ lib/Pages/PageAllRating.php
@@ -10,14 +10,12 @@ namespace SebLucas\Cops\Pages;
 
 use SebLucas\Cops\Calibre\Rating;
 
-use function SebLucas\Cops\Language\localize;
-
 class PageAllRating extends Page
 {
     public function InitializeContent()
     {
         $this->title = localize("ratings.title");
-        $this->entryArray = Rating::getAllRatings();
+        $this->entryArray = Rating::getAllRatings($this->getDatabaseId());
         $this->idPage = Rating::PAGE_ID;
     }
 }