+++ lib/Pages/PageSerieDetail.php
@@ -15,9 +15,9 @@ class PageSerieDetail extends Page
 {
     public function InitializeContent()
     {
-        $serie = Serie::getSerieById($this->idGet);
+        $serie = Serie::getSerieById($this->idGet, $this->getDatabaseId());
         $this->title = $serie->name;
-        [$this->entryArray, $this->totalNumber] = Book::getBooksBySeries($this->idGet, $this->n);
+        [$this->entryArray, $this->totalNumber] = Book::getBooksBySeries($this->idGet, $this->n, $this->getDatabaseId());
         $this->idPage = $serie->getEntryId();
     }
 }